feat(wxapi): 新增视频号小店运费模板相关接口

+++ src/SKIT.FlurlHttpClient.Wechat.Api/Models/Channels/ECCategory/ChannelsECCategoryAvailableSonCategoriesGetRequest.cs
@@ -10,6 +10,6 @@
         /// </summary>
         [Newtonsoft.Json.JsonProperty("f_cat_id")]
         [System.Text.Json.Serialization.JsonPropertyName("f_cat_id")]
-        public int ParentCategoryId { get; set; }
+        public long ParentCategoryId { get; set; }
     }
 }
